+++ api/providers/inventory/inventory_provider.py
@@ -6,6 +6,9 @@ from api.external import lta, inventory, lpdaac, nlaps
 from api import InventoryException, InventoryConnectionException
 from api.domain import sensor
 from api.system.logger import ilogger as logger
+from api.providers.configuration.configuration_provider import ConfigurationProvider
+
+config = ConfigurationProvider()
 
 
 class InventoryProviderV0(InventoryInterfaceV0):
@@ -34,21 +37,21 @@ class InventoryProviderV0(InventoryInterfaceV0):
                 lpdaac_ls.extend(order[key]['inputs'])
 
         if lta_ls:
-            if 'LANDSAT' in os.getenv('ESPA_M2M_MODE', '') and inventory.available():
+            if config.is_m2m_val_enabled and inventory.available():
                 results.update(self.check_dmid(lta_ls, contactid))
-            else:
-                if not lta.check_lta_available():
-                    msg = 'Could not connect to LTA data source'
-                    raise InventoryConnectionException(msg)
+            elif lta.check_lta_available():
                 results.update(self.check_LTA(lta_ls))
-        if lpdaac_ls:
-            if 'MODIS' in os.getenv('ESPA_M2M_MODE', '') and inventory.available():
-                results.update(self.check_dmid(lpdaac_ls, contactid))
             else:
-                if not lpdaac.check_lpdaac_available():
-                    msg = 'Could not connect to LPDAAC data source'
-                    raise InventoryConnectionException(msg)
+                msg = 'Could not connect to Landsat data source'
+                raise InventoryConnectionException(msg)
+        if lpdaac_ls:
+            if config.is_m2m_val_enabled and inventory.available():
+                results.update(self.check_dmid(lpdaac_ls, contactid))
+            elif lpdaac.check_lpdaac_available():
                 results.update(self.check_LPDAAC(lpdaac_ls))
+            else:
+                msg = 'Could not connect to MODIS data source'
+                raise InventoryConnectionException(msg)
 
         not_avail = []
         for key, val in results.items():
@@ -64,21 +67,15 @@ class InventoryProviderV0(InventoryInterfaceV0):
         not_avail = nlaps.products_are_nlaps(prod_ls)
         if not_avail:
             raise InventoryException(not_avail)
-        logger.info('@@ VERIFY: FETCH via MACHINE-TO-MACHINE')
         token = inventory.get_cached_session()
-        results = inventory.get_cached_verify_scenes(token, prod_ls)
-        logger.info('@@ VERIFY: COMPLETE via MACHINE-TO-MACHINE')
-        return results
+        return inventory.get_cached_verify_scenes(token, prod_ls)
 
     @staticmethod
     def check_LTA(prod_ls):
         not_avail = nlaps.products_are_nlaps(prod_ls)
         if not_avail:
             raise InventoryException(not_avail)
-        logger.info('@@ VERIFY: FETCH via OrderWrapperService')
-        results = lta.verify_scenes(prod_ls)
-        logger.info('@@ VERIFY: COMPLETE via OrderWrapperService')
-        return results
+        return lta.verify_scenes(prod_ls)
 
     @staticmethod
     def check_LPDAAC(prod_ls):